Prevent privilege escalations through invalid JWTs
Louketo has a bug revolving around this PR: https://github.com/coreos/go-oidc/pull/97
essentially, the JWT verification function first checks the claims of the JWT are valid (including expiry times), then checks the signature of the token is valid.
Louketo checks the response of this and returns a custom error if the token has expired to allow the auth middleware to use a refresh token to regenerate an access token.
Unfortunately, the access token it generates uses the user context for the incoming request. This user context is parsed from the access token JWT but is not verified, meaning it is blindly trusted.
Louketo will then regenerate a correct accessToken from Keycloak, which will be valid and have the correct claim/scopes as generated in Keycloak. I think the issue is that for the rest of the request, it has this new token in the context but still has the leftover rubbish user context parsed from the fake user we injected earlier.